Fix call to MallTransaction#refund
It is not possible to call the refund() method of the MallTransaction
class because it is expecting a variable called child_buy_order to be
defined, yet this variable is not available anywhere (and it is not
used).

Unfortunately it seems that MallTransaction is not tested, so I cannot
add a test case to the test suite, but calling this method in a
bin/console yields the expected results in terms of API communications
after doing a refund and validating it with status().

At the same time, I am adding a validation for the token because it
seems that it is missing and it is probably a good idea given that every
other method in MallTransaction, plus methods in Transaction, do
validate the length of the token.